+++ lib/appium_lib_core/driver.rb
@@ -228,16 +228,7 @@ module Appium
         @driver.manage.timeouts.implicit_wait = wait
       rescue ::Selenium::WebDriver::Error::UnknownError => e
         unless e.message.include?('The operation requested is not yet implemented')
-          raise e.message, ::Appium::Core::Error::ServerError
-        end
-
-        Appium::Logger.debug(e.message)
-        {}
-      rescue ::Selenium::WebDriver::Error::WebDriverError => e
-        # FIXME: Temporary rescue until Appium support W3C's implicit wait
-        # https://github.com/jlipps/simple-wd-spec#set-timeouts
-        unless e.message.include?('Parameters were incorrect. We wanted {"required":["type","ms"]} and you sent ["implicit"]')
-          raise e.message, ::Appium::Core::Error::ServerError
+          raise ::Appium::Core::Error::ServerError, e.message
         end
 
         Appium::Logger.debug(e.message)